Add information to the project repo
This heavily modifies Code for Baltimore's project template, making it
suitable for this project.

Besides directly adding information about this project, this
also heavily simplifies the included documentation. This is either due
to it being irrelevant at our current stage of the project, or because I
feel it is unnecessary. This documentation may be readded once this
project get to the relevant stage, or when specifying it is useful.